feat: add Simplified Chinese
Added Simplified Chinese support, using two-letter locale code according to README.md. However, it should be noted that this will result in Traditional Chinese and even various regions of Simplified Chinese and Traditional Chinese not being represented. My suggestion is to use zh_CN and zh_TW and so on with regional abbreviations after the two-letter locale code to distinguish them.